refactor: disable sentry
