+++ src/app/common/hooks/analytics/use-analytics.ts
@@ -1,7 +1,7 @@
 import { useCurrentNetworkState } from '@app/store/network/networks.hooks';
 import { useLocation } from 'react-router-dom';
 import { EventParams, PageParams } from '@segment/analytics-next/dist/pkg/core/arguments-resolver';
-import { analytics } from '@app/common/segment-init';
+import { analytics } from '@shared/utils/analytics';
 import { logger } from '@shared/logger';
 import { useWalletType } from '@app/common/use-wallet-type';
 import { IS_TEST_ENV } from '@shared/environment';
