+++ src/app/pages/onboarding/welcome/welcome.layout.tsx
@@ -9,7 +9,7 @@ import { PrimaryButton } from '@app/components/primary-button';
 import { CenteredPageContainer } from '@app/components/centered-page-container';
 import ExploreStacks from '@assets/images/onboarding/explore-stacks.png';
 import { OnboardingSelectors } from '@tests/integration/onboarding/onboarding.selectors';
-import { isLedgerFeatureEnabled } from '@shared/constants';
+import { featureFlags } from '@shared/feature-flags';
 
 const WelcomeIllustration = () => (
   <Box
@@ -92,7 +92,7 @@ export function WelcomeLayout(props: WelcomeLayoutProps): JSX.Element {
               >
                 Sign in with Secret Key
               </Link>
-              {isLedgerFeatureEnabled && (
+              {featureFlags.ledgerEnabled && (
                 <>
                   {' '}
                   or{' '}
