+++ src/shared/constants.ts
@@ -24,8 +24,6 @@ export const MICROBLOCKS_ENABLED = !IS_TEST_ENV && true;
 export const GITHUB_ORG = 'hirosystems';
 export const GITHUB_REPO = 'stacks-wallet-web';
 
-export const isLedgerFeatureEnabled = true;
-
 export interface Network {
   url: string;
   name: string;
