+++ src/shared/feature-flags.ts
@@ -1,0 +1,3 @@
+export const featureFlags = {
+  ledgerEnabled: true,
+};
